refactoring

+++ rentACar/src/main/java/com/etiya/rentACar/business/concretes/ColorManager.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import com.etiya.rentACar.business.constants.messages.ColorMessages;
+import com.etiya.rentACar.core.utilities.business.BusinessRules;
+import com.etiya.rentACar.core.utilities.results.ErrorResult;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,10 +22,10 @@ import com.etiya.rentACar.entities.Color;
 
 @Service
 public class ColorManager implements ColorService{
-	
+
 	private ColorDao colorDao;
 	private ModelMapperService modelMapperService;
-	
+
 	@Autowired
 	public ColorManager(ColorDao colorDao, ModelMapperService modelMapperService) {
 		super();
@@ -34,17 +36,21 @@ public class ColorManager implements ColorService{
 
 	@Override
 	public List<ColorSearchListDto> getColors() {
-		
+
 		List<Color> result = this.colorDao.findAll();
 		List<ColorSearchListDto> response = result.stream().map(color->modelMapperService.forDto()
 				.map(color, ColorSearchListDto.class)).collect(Collectors.toList());
-				
+
 		return response;
 	}
 
 
 	@Override
 	public Result save(CreateColorRequest createColorRequest) {
+		Result result = BusinessRules.run(checkExistingColor(createColorRequest.getColorName()));
+		if (result != null){
+			return result;
+		}
 		Color color = modelMapperService.forRequest().map(createColorRequest, Color.class);
 		this.colorDao.save(color);
 		return new SuccessResult(ColorMessages.add);
@@ -53,6 +59,10 @@ public class ColorManager implements ColorService{
 
 	@Override
 	public Result delete(DeleteColorRequest deleteColorRequest) {
+		Result result = BusinessRules.run(checkExistingColorId(deleteColorRequest.getColorId()));
+		if (result != null){
+			return result;
+		}
 		Color color = modelMapperService.forRequest().map(deleteColorRequest, Color.class);
 		this.colorDao.delete(color);
 		return new SuccessResult(ColorMessages.delete);
@@ -61,8 +71,29 @@ public class ColorManager implements ColorService{
 
 	@Override
 	public Result update(UpdateColorRequest updateColorRequest) {
+		Result result = BusinessRules.run(checkExistingColor(updateColorRequest.getColorName()), checkExistingColorId(updateColorRequest.getColorId()));
+		if (result != null){
+			return result;
+		}
 		Color color = modelMapperService.forRequest().map(updateColorRequest, Color.class);
 		this.colorDao.save(color);
 		return new SuccessResult(ColorMessages.update);
 	}
-}
+
+	private Result checkExistingColor(String colorName1) {
+		String lowerCaseColorName = colorName1.toLowerCase();
+		for (Color color : colorDao.findAll()) {
+			String lowerCaseExistingColorName = color.getColorName().toLowerCase();
+			if(lowerCaseColorName.equals(lowerCaseExistingColorName)) {
+				return new ErrorResult(ColorMessages.duplicationError);
+			}
+		}
+		return new SuccessResult();
+	}
+
+	private Result checkExistingColorId(int colorId){
+		if (colorDao.existsById(colorId)){
+			return new SuccessResult();
+		}
+		return new ErrorResult(ColorMessages.colorIdNotFound);
+	}}
